Serve metrics server in https
