+++ test/e2e/bindata/assets/04_deployment.yaml
@@ -45,6 +45,8 @@ spec:
           ports:
             - containerPort: 9449
               protocol: TCP
+            - containerPort: 8443
+              protocol: TCP
           volumeMounts:
             - mountPath: "/etc/secrets"
               name: certs-dir
