+++ test/e2e/bindata/assets/06_service.yaml
@@ -20,3 +20,25 @@ spec:
     app: openshift-cli-manager
   sessionAffinity: None
   type: ClusterIP
+---
+apiVersion: v1
+kind: Service
+metadata:
+  annotations:
+    include.release.openshift.io/self-managed-high-availability: "true"
+    include.release.openshift.io/single-node-developer: "true"
+    service.beta.openshift.io/serving-cert-secret-name: openshift-cli-manager-serving-cert
+    exclude.release.openshift.io/internal-openshift-hosted: "true"
+  labels:
+    app: openshift-cli-manager
+  name: openshift-cli-manager-metrics
+  namespace: openshift-cli-manager-operator
+spec:
+  clusterIP: None
+  ports:
+    - port: 8443
+      protocol: TCP
+      targetPort: 8443
+      name: cli-manager-metrics-port
+  selector:
+    app: openshift-cli-manager
